Add cucumber test Gulp task
Add a cucumber test Gulp task.
Add a task for setting environment variables so that the same port can be passed to server:start and picked up by the tests.
Note: this work is dependent on the acceptance on a pull request in gulp-cucumber https://github.com/vgamula/gulp-cucumber/pull/16