Improving the performance of retrieving specifications
When there is a lot of files, there is a performance hit in the
addEditLinks function by not caching the template function. This
commit adds caching on the project level which shaves off 30%
of the rendering time.

Also improvement to select branch drop down as with long branch names
it is not wide enough to read full branch name.